# Improve bulk file download performance