--- a/apps/prairielearn/src/pages/instructorAssessmentDownloads/instructorAssessmentDownloads.sql
+++ b/apps/prairielearn/src/pages/instructorAssessmentDownloads/instructorAssessmentDownloads.sql
@@ -159,37 +159,51 @@ WITH
       g.id,
       u.user_id,
       ai.number DESC
+  ),
+  final_submissions AS (
+    SELECT DISTINCT
+      ON (ai.id, q.qid) u.uid,
+      u.uin,
+      q.qid,
+      iq.score_perc AS old_score_perc,
+      iq.auto_points AS old_auto_points,
+      iq.manual_points AS old_manual_points,
+      aq.max_points,
+      aq.max_auto_points,
+      aq.max_manual_points,
+      s.feedback AS old_feedback,
+      s.id AS submission_id,
+      s.params,
+      s.true_answer,
+      CASE
+        WHEN $include_files THEN s.submitted_answer
+        ELSE (s.submitted_answer - '_files')
+      END AS submitted_answer,
+      s.partial_scores AS old_partial_scores,
+      ai.group_name,
+      ai.uid_list
+    FROM
+      submissions AS s
+      JOIN variants AS v ON (v.id = s.variant_id)
+      JOIN instance_questions AS iq ON (iq.id = v.instance_question_id)
+      JOIN final_assessment_instances AS ai ON (ai.id = iq.assessment_instance_id)
+      LEFT JOIN users AS u ON (u.user_id = ai.user_id)
+      JOIN assessment_questions AS aq ON (aq.id = iq.assessment_question_id)
+      JOIN questions AS q ON (q.id = aq.question_id)
+    ORDER BY
+      ai.id,
+      q.qid,
+      s.date DESC
   )
-SELECT DISTINCT
-  ON (ai.id, q.qid) u.uid,
-  u.uin,
-  q.qid,
-  iq.score_perc AS old_score_perc,
-  iq.auto_points AS old_auto_points,
-  iq.manual_points AS old_manual_points,
-  aq.max_points,
-  aq.max_auto_points,
-  aq.max_manual_points,
-  s.feedback AS old_feedback,
-  s.id AS submission_id,
-  s.params,
-  s.true_answer,
-  (s.submitted_answer - '_files') AS submitted_answer,
-  s.partial_scores AS old_partial_scores,
-  ai.group_name,
-  ai.uid_list
+SELECT
+  *
 FROM
-  submissions AS s
-  JOIN variants AS v ON (v.id = s.variant_id)
-  JOIN instance_questions AS iq ON (iq.id = v.instance_question_id)
-  JOIN final_assessment_instances AS ai ON (ai.id = iq.assessment_instance_id)
-  LEFT JOIN users AS u ON (u.user_id = ai.user_id)
-  JOIN assessment_questions AS aq ON (aq.id = iq.assessment_question_id)
-  JOIN questions AS q ON (q.id = aq.question_id)
+  final_submissions
 ORDER BY
-  ai.id,
-  q.qid,
-  s.date DESC;
+  uid,
+  group_name,
+  qid,
+  submission_id;
 
 -- BLOCK assessment_instance_submissions
 WITH
@@ -232,6 +246,12 @@ WITH
         ELSE NULL
       END AS correct,
       s.feedback,
+      row_number() OVER (
+        PARTITION BY
+          v.id
+        ORDER BY
+          s.date
+      )::integer AS submission_number,
       (
         row_number() OVER (
           PARTITION BY
@@ -300,257 +320,6 @@ ORDER BY
   assessment_instance_number,
   qid,
   instance_question_number,
-  variant_number,
-  date;
-
--- BLOCK files_for_manual_grading
-WITH
-  final_assessment_instances AS (
-    SELECT DISTINCT
-      ON (g.id, u.user_id) u.user_id,
-      g.id AS group_id,
-      ai.id,
-      g.name AS group_name
-    FROM
-      assessment_instances AS ai
-      LEFT JOIN groups AS g ON (g.id = ai.group_id)
-      LEFT JOIN users AS u ON (u.user_id = ai.user_id)
-    WHERE
-      ai.assessment_id = $assessment_id
-    ORDER BY
-      g.id,
-      u.user_id,
-      ai.number DESC
-  ),
-  submissions_with_files AS (
-    SELECT DISTINCT
-      ON (ai.id, q.qid) u.uid,
-      u.uin,
-      q.qid,
-      s.id AS submission_id,
-      s.submitted_answer,
-      v.params,
-      ai.group_name
-    FROM
-      submissions AS s
-      JOIN variants AS v ON (v.id = s.variant_id)
-      JOIN instance_questions AS iq ON (iq.id = v.instance_question_id)
-      JOIN final_assessment_instances AS ai ON (ai.id = iq.assessment_instance_id)
-      LEFT JOIN users AS u ON (u.user_id = ai.user_id)
-      JOIN assessment_questions AS aq ON (aq.id = iq.assessment_question_id)
-      JOIN questions AS q ON (q.id = aq.question_id)
-    WHERE
-      (
-        v.params ? 'fileName'
-        AND s.submitted_answer ? 'fileData'
-      )
-      OR (s.submitted_answer ? '_files')
-    ORDER BY
-      ai.id,
-      q.qid,
-      s.date DESC
-  ),
-  all_files AS (
-    SELECT
-      uid,
-      group_name,
-      uin,
-      qid,
-      submission_id,
-      (
-        CASE
-          WHEN submitted_answer ? 'fileData' THEN params ->> 'fileName'
-          WHEN submitted_answer ? '_files' THEN f.file ->> 'name'
-        END
-      ) as filename,
-      (
-        CASE
-          WHEN submitted_answer ? 'fileData' THEN submitted_answer ->> 'fileData'
-          WHEN submitted_answer ? '_files' THEN f.file ->> 'contents'
-        END
-      ) as contents
-    FROM
-      submissions_with_files AS s
-      LEFT JOIN (
-        SELECT
-          submission_id AS id,
-          jsonb_array_elements(submitted_answer -> '_files') AS file
-        FROM
-          submissions_with_files
-        WHERE
-          submitted_answer ? '_files'
-      ) f ON (f.id = submission_id)
-  )
-SELECT
-  (
-    (
-      CASE
-        WHEN $group_work THEN group_name
-        ELSE uid || '_' || uin
-      END
-    ) || '_' || qid || '_' || submission_id || '_' || filename
-  ) AS filename,
-  contents
-FROM
-  all_files
-WHERE
-  filename IS NOT NULL
-  AND contents IS NOT NULL
-ORDER BY
-  (
-    CASE
-      WHEN $group_work THEN group_name
-      ELSE uid
-    END
-  ),
-  (
-    CASE
-      WHEN $group_work THEN NULL
-      ELSE uin
-    END
-  ),
-  qid,
-  filename,
-  submission_id;
-
--- BLOCK assessment_instance_files
-WITH
-  all_submissions_with_files AS (
-    SELECT
-      s.id AS submission_id,
-      u.uid,
-      ai.number AS assessment_instance_number,
-      q.qid,
-      v.number AS variant_number,
-      v.params,
-      s.date,
-      s.submitted_answer,
-      row_number() OVER (
-        PARTITION BY
-          v.id
-        ORDER BY
-          s.date
-      ) AS submission_number,
-      (
-        row_number() OVER (
-          PARTITION BY
-            v.id
-          ORDER BY
-            s.date DESC,
-            s.id DESC
-        )
-      ) = 1 AS final_submission_per_variant,
-      (
-        row_number() OVER (
-          PARTITION BY
-            v.id
-          ORDER BY
-            s.score DESC NULLS LAST,
-            s.date DESC,
-            s.id DESC
-        )
-      ) = 1 AS best_submission_per_variant,
-      g.name AS group_name
-    FROM
-      assessments AS a
-      JOIN assessment_instances AS ai ON (ai.assessment_id = a.id)
-      LEFT JOIN group_configs AS gc ON (gc.assessment_id = a.id)
-      LEFT JOIN groups AS g ON (
-        g.id = ai.group_id
-        AND g.group_config_id = gc.id
-      )
-      LEFT JOIN users AS u ON (u.user_id = ai.user_id)
-      JOIN instance_questions AS iq ON (iq.assessment_instance_id = ai.id)
-      JOIN assessment_questions AS aq ON (aq.id = iq.assessment_question_id)
-      JOIN questions AS q ON (q.id = aq.question_id)
-      JOIN variants AS v ON (v.instance_question_id = iq.id)
-      JOIN submissions AS s ON (s.variant_id = v.id)
-    WHERE
-      a.id = $assessment_id
-      AND (
-        (
-          v.params ? 'fileName'
-          AND s.submitted_answer ? 'fileData'
-        )
-        OR (s.submitted_answer ? '_files')
-      )
-  ),
-  use_submissions_with_files AS (
-    SELECT
-      *
-    FROM
-      all_submissions_with_files
-    WHERE
-      (
-        $include_all
-        OR (
-          $include_final
-          AND final_submission_per_variant
-        )
-        OR (
-          $include_best
-          AND best_submission_per_variant
-        )
-      )
-  ),
-  all_files AS (
-    SELECT
-      uid,
-      group_name,
-      assessment_instance_number,
-      qid,
-      variant_number,
-      date,
-      submission_id,
-      submission_number,
-      (
-        CASE
-          WHEN submitted_answer ? 'fileData' THEN params ->> 'fileName'
-          WHEN submitted_answer ? '_files' THEN f.file ->> 'name'
-        END
-      ) as filename,
-      (
-        CASE
-          WHEN submitted_answer ? 'fileData' THEN submitted_answer ->> 'fileData'
-          WHEN submitted_answer ? '_files' THEN f.file ->> 'contents'
-        END
-      ) as contents
-    FROM
-      use_submissions_with_files AS s
-      LEFT JOIN (
-        SELECT
-          submission_id AS id,
-          jsonb_array_elements(submitted_answer -> '_files') AS file
-        FROM
-          use_submissions_with_files
-        WHERE
-          submitted_answer ? '_files'
-      ) f ON (f.id = submission_id)
-  )
-SELECT
-  (
-    (
-      CASE
-        WHEN $group_work THEN group_name
-        ELSE uid
-      END
-    ) || '_' || assessment_instance_number || '_' || qid || '_' || variant_number || '_' || submission_number || '_' || submission_id || '_' || filename
-  ) AS filename,
-  contents
-FROM
-  all_files
-WHERE
-  filename IS NOT NULL
-  AND contents IS NOT NULL
-ORDER BY
-  (
-    CASE
-      WHEN $group_work THEN group_name
-      ELSE uid
-    END
-  ),
-  assessment_instance_number,
-  qid,
   variant_number,
   date;