This PR improves the performance of bulk file downloads for assessments, particularly the `*_all_files.zip`, `*_final_files.zip`, `*_best_files.zip`, and `*_files_for_manual_grading.zip` archives.

We got a report recently that an instructor was unable to download `*_best_files.zip` for their assignment. We believe this is due to the large amount of data in play: there were ~500 students, each submission was ~1-2MB in size, and most students submitted many times. This resulted in probably dozens of gigabytes of data involved. Specifically, as part of the `assessment_instance_files`, we'd end up reading `submissions.submitted_answers` for _every single submission_, not just the best one. This was a huge amount of data to page in and scan through. Even then, the query ran much longer than we'd have expected it to: in testing, it ran for over 5 minutes before I terminated it.

@mwest1066 and I spent some time looking at query plans and trying to optimize the existing query. Ultimately, we weren't able to determine exactly why things were so slow, but we believe it has something to do with accessing `submitted_answer` in a `CASE` statement. We observed that the all/final/best submission CSVs started downloading much faster, and so we made the decision to just reuse the same backing queries for the file queries. This means we need to push the extraction of files into JavaScript, which is entirely acceptable.

An important observation we made during our investigation is that the really important thing for huge downloads is to start sending bytes as soon as possible. In production, our load balancer is configured with an idle connection timeout of 60 seconds. This means we have 60 seconds to start sending data before the connection is terminated. What this means in practice is that we have 60 seconds for the database query to start giving us data. We use a cursor to read data so we don't have to buffer it all into memory, we just have to get the first set of rows from the cursor.

In our testing, using the existing `assessment_instance_submissions` and `submissions_for_manual_grading` queries, we were able to start returning data in ~2 seconds, instead of many minutes. The file itself takes a while to download, but it won't time out!