@echuber2 2024-12-04
Does this address this old issue? (Or was it already addressed?) https://github.com/PrairieLearn/PrairieLearn/issues/4982